requirements.txt
Werkzeug==1.0.1
pymongo~=3.11.2
jsonschema~=3.2.0
Flask~=1.1.2
python-dotenv~=0.15.0
pytest